Use Vec<Vec<Entry<T>>> for slots in PinSlab
The slots were already Vecs but stored as pointers. This change makes
them Vecs which allows many of the uses of `unsafe` to be removed.

This could potentially come with a performance decrease due to bounds
checking, but in running `cargo bench` it does not seem to have made
much difference.