Use Arc for Internals waker
Using an Arc instead prevents an allocation on every clone of the waker.
Instead, we only need the allocation when converting from a RefWaker
to Internals. We still end up with the same number of atomic ref count
changes since before we had to increment the count on the Arc<Shared>
but that is no longer necessary.