Cache wakers in shared task structure
This allows wakers to be reused when upgrading a RefWaker to an
InternalWaker.

This is probably sub-optimal, but it serves as a starting point for a
better one. This change adds some new locking overhead, and because we
store weak pointers to the child wakers that end up having to be
reallocated often. Future changes will store them inline and pass
pointers around.